pyproject.toml
# pyproject.toml (Versione FitManager 3.0 - Agenda Update)

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fit-manager-ai"
version = "3.0.0" # Aggiornato alla v3
authors = [
  { name="G. Verardo", email="tuo.email@example.com" },
]
description = "Piattaforma AI per la gestione di studi fitness, atleti e nutrizione."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT License" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Framework :: Streamlit",
]
dependencies = [
    # --- CORE LIBS (Base) ---
    "streamlit>=1.54.0",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "requests",
    "openpyxl==3.1.2",
    "python-dotenv==1.0.1",
    "plotly==5.22.0",
    "feedparser",
    "streamlit-calendar", # <--- NUOVA DIPENDENZA FONDAMENTALE
    "watchdog",           # <--- Per gestione file system
    "pydantic>=2.7.4",    # <--- REPOSITORY PATTERN
    "python-dateutil",    # <--- Dependency per repository pattern
    "python-docx==1.1.2", # <--- Trainer DNA: import schede Word
    "cryptography",       # <--- Backup crittografato (.fitbackup)

    # --- API LAYER (FastAPI + Auth) ---
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",           # <--- ORM (SQLAlchemy + Pydantic) — database-agnostic
    "python-jose[cryptography]",  # <--- JWT tokens
    "bcrypt",             # <--- Password hashing
    "python-multipart",   # <--- Form data per auth

    # --- LANGCHAIN ECOSYSTEM (AI Logic) ---
    "langchain==0.2.5",
    "langchain-community",
    "langchain-core",
    "langchain-chroma",
    "langchain-ollama",
    "langchain-text-splitters",

    # --- RAG & ML (Motore AI Locale) ---
    "ollama==0.5.4",
    "PyMuPDF==1.24.1",
    "chromadb==0.4.24",
    "torch==2.3.0",
    "transformers==4.41.2",
    "sentence-transformers==2.7.0",
    "scikit-learn==1.5.0",
    "joblib==1.4.2",
]

[project.urls]
"Homepage" = "https://github.com/tuo-user/fit-manager-ai"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "server*", "api*", "tools*", "knowledge_base*"]
exclude = ["data*", "venv*"]